smol refactor
