+++ CreeDictionary/CreeDictionary/hfstol.py
@@ -11,7 +11,7 @@ from subprocess import DEVNULL, check_output
 from tempfile import TemporaryFile
 from typing import IO, Generator, Iterable, List, Tuple
 
-from constants import Analysis
+from constants.data_classes import Analysis
 from utils.shared_res_dir import shared_res_dir as res
 
 # Ensure we can load everything!
