+++ CreeDictionary/CreeDictionary/templatetags/creedictionary_extras.py
@@ -5,7 +5,8 @@
 Template tags related to the Cree Dictionary specifically.
 """
 
-from constants import DEFAULT_ORTHOGRAPHY, ORTHOGRAPHY_NAME
+from constants.vars import DEFAULT_ORTHOGRAPHY
+from constants.data_classes import ORTHOGRAPHY_NAME
 from cree_sro_syllabics import sro2syllabics
 from CreeDictionary.utils import url_for_query
 from django import template
