+++ CreeDictionary/CreeDictionary/views.py
@@ -1,7 +1,8 @@
 from http import HTTPStatus
 
 from API.models import Wordform
-from constants import ORTHOGRAPHY_NAME, ParadigmSize
+from constants import ParadigmSize
+from constants.data_classes import ORTHOGRAPHY_NAME
 from CreeDictionary.forms import WordSearchForm
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
