+++ CreeDictionary/DatabaseManager/xml_importer.py
@@ -14,7 +14,7 @@ from DatabaseManager import xml_entry_lemma_finder
 from DatabaseManager.cree_inflection_generator import expand_inflections
 from DatabaseManager.log import DatabaseManagerLogger
 from DatabaseManager.xml_consistency_checker import does_lc_match_xml_entry
-from constants import POS
+from constants.enums import POS
 from utils import fst_analysis_parser
 from utils.crkeng_xml_utils import extract_l_str, parse_xml_lc
