+++ CreeDictionary/constants/types.py
@@ -1,0 +1,7 @@
+from typing import NewType
+
+# analysis but concatenated
+Label = NewType("Label", str)
+FSTTag = NewType("FSTTag", str)
+FSTLemma = NewType("FSTLemma", str)
+ConcatAnalysis = NewType("ConcatAnalysis", str)
