+++ CreeDictionary/constants/vars.py
@@ -1,0 +1,1 @@
+DEFAULT_ORTHOGRAPHY = "Latn"
